Work in Progress - Applying Statistics to Chores
Statistics is everywhere, from cycling, to studying, to working, to even daily chores. This is a data story of my own experience of visualizing Hypothesis Testing in mt daily life. The images are from Google, but the struggle and concept application is mine. 
It is still a work in progress and I update it frequently.